requirements.txt
torch>=2.1.0
torchvision>=0.16.0
lightning>=2.1.0
einops>=0.7.0
lpips>=0.1.4
omegaconf>=2.3.0
tensorboard>=2.15.0
wandb>=0.16.0
Pillow>=10.0.0
numpy>=1.24.0
